Multiplex everything over a single connection.
Instead of maintaining a control connection and additional data
transfer connections, try to multiplex everything over a single TCP
connection. While this incurs overhead it simplifies the protocol:

Once the agent is connected to the gateway, opening a new stream
for data transfer is fast because the agent does not need to open a
new TCP connection, perform another TLS handshake and authenticate
and authorise itself to a potentially different gateway node. There
is no need to report back the allocated address, no need to sign
authorisation tokens and to encrypt address information.

In addition if there is a transfer error local to the stream the
transfer may be retried without having to re-establish the TCP
connection between agent and gateway again.