Add connect option `use_half_close` to protocol.
If true, the agent will attempt to shutdown the write-half of the
socket to its upstream system (e.g. a database). The Yamux stream
will always be shutdown, regardless of the setting.

NB that up to this commit we always performed a half-close. What
is new is that we offer the option to not do this which is useful
for TLS traffic that is not on version 1.3.