Add `Error` and `Accepted` to protocol.
A server error contains a message that the agent will log.
The `Server::Accepted` command is issued by the server to signal to the
agent that it has accepted the connection. Its primary use is to ease
the exponential backoff handling. Currently the agent does not know
when it can safely assume that the server will keep it. The exponential
backoff only kicks in when connection establishment on TCP level fails
but after that the retries are fairly constant. With this change the
agent will only reset its counter when the `Accepted` command has been
received. Disconnects prior to that will handled with the same backoff
mechanism as TCP or TLS connection errors.