Trace online status when processing messages.
Also skip branch if connection.inbound is ready but the agent is offline.
Especially when the connection is closed, connection.inbound is always
ready. We let tokio::select handle the choice between multiple branches
which it does randomly. Since we do not expect many pending stream
messages we do not replace the `connection.inbound.recv()` with another
future. And if polled too often the budget of the future will quickly be
used up.