Use cross-rs/cross for compiling linux musl targets.
This avoids linking issues of the previous build setup after
switching to the latest rustls and the new default
aws-lc-rs crypto provider.

The windows build is updated to install NASM required for the
compilation of aws-lc-rs on the x86_64-pc-windows-msvc target.

The MacOS build is updated to now build on and for >= MacOS 12
as MacOS 11 is longer supported both by Apple and Github CI.

Stripping of binaries is now left to cargo/rustc.